+++ app/src/main/java/eu/kanade/tachiyomi/data/source/online/english/Mangafox.kt
@@ -11,7 +11,7 @@ import java.text.ParseException
 import java.text.SimpleDateFormat
 import java.util.*
 
-class Mangafox(override val id: Int) : ParsedOnlineSource() {
+class Mangafox(override val id: Long) : ParsedOnlineSource() {
 
     override val name = "Mangafox"
