+++ app/src/main/java/eu/kanade/tachiyomi/data/source/online/english/Mangasee.kt
@@ -16,7 +16,7 @@ import org.jsoup.nodes.Element
 import java.text.SimpleDateFormat
 import java.util.regex.Pattern
 
-class Mangasee(override val id: Int) : ParsedOnlineSource() {
+class Mangasee(override val id: Long) : ParsedOnlineSource() {
 
     override val name = "Mangasee"
