+++ app/src/main/java/eu/kanade/tachiyomi/data/source/online/russian/Mangachan.kt
@@ -12,7 +12,7 @@ import org.jsoup.nodes.Element
 import java.text.SimpleDateFormat
 import java.util.*
 
-class Mangachan(override val id: Int) : ParsedOnlineSource() {
+class Mangachan(override val id: Long) : ParsedOnlineSource() {
 
     override val name = "Mangachan"
